+++ train-interfaces/src/main/java/hu/bme/mit/train/interfaces/TrainUser.java
@@ -2,10 +2,10 @@ package hu.bme.mit.train.interfaces;
 
 public interface TrainUser {
 
-	int getJoystickPosition();
+	double getJoystickPosition();
 
 	boolean getAlarmFlag();
 
-	void overrideJoystickPosition(int joystickPosition);
+	void overrideJoystickPosition(double joystickPosition);
 
 }
